+++ src/app/pages/angular/angular-routing.module.ts
@@ -12,6 +12,7 @@ const routes: Routes = [
       { path: 'ng-common', loadChildren: () => import('./ng-common/ng-common.module').then((m) => m.NgCommonModule) },
       { path: 'ng-core', loadChildren: () => import('./ng-core/ng-core.module').then((m) => m.NgCoreModule) },
       { path: 'ng-form', loadChildren: () => import('./ng-form/ng-form.module').then((m) => m.NgFormModule) },
+      { path: 'ng-route', loadChildren: () => import('./ng-route/ng-route.module').then((m) => m.NgRouteModule) },
       { path: 'ng-animations', loadChildren: () => import('./ng-animations/ng-animations.module').then((m) => m.NgAnimationsModule) },
 
       { path: '', redirectTo: 'life-cycle', pathMatch: 'full' },