+++ src/app/pages/angular/ng-route/nest-router/child1-router.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-child1-router',
+  template: '<p style="color: #0f0;">child1-router works!</p>',
+})
+export class Child1RouterComponent {
+
+}