+++ src/app/pages/angular/ng-route/nest-router/child2-router.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-child2-router',
+  template: '<p style="color: #f0f;">child2-router works!</p>',
+})
+export class Child2RouterComponent {
+
+}