+++ src/app/pages/angular/ng-route/nest-router/nest-router/nest-router.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-nest-router',
+  templateUrl: './nest-router.component.html',
+  styleUrls: ['./nest-router.component.scss']
+})
+export class NestRouterComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}