+++ src/app/pages/angular/ng-route/refresh-candeactivate/refresh-candeactivate.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+
+import { RefreshCandeactivatePageRoutingModule } from './refresh-candeactivate-routing.module';
+
+import { RefreshCandeactivatePage } from './refresh-candeactivate.page';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FormsModule,
+    RefreshCandeactivatePageRoutingModule
+  ],
+  declarations: [RefreshCandeactivatePage]
+})
+export class RefreshCandeactivatePageModule {}