+++ ChangeLog.md
@@ -15,3 +15,6 @@
 
 ### 2024-05-15
 1. 完成迁移 angular - @angular/forms 内容
+
+### 2024-05-15
+1. 完成迁移 angular - @angular/router 内容